Added the latest device page, slug page, sidebar component, and table component
