+++ app/(client)/layout.js
@@ -34,9 +34,7 @@ export default function RootLayout({ children }) {
           enableSystem={true}
         >
           <Navbar />
-          <main className="max-w-6xl mx-auto">
-            {children}
-          </main>
+          <main className="max-w-6xl mx-auto">{children}</main>
         </ThemeProvider>
       </body>
     </html>
