+++ app/phones.json
@@ -7,7 +7,8 @@
       "camera": "108MP + 12MP Ultra-wide + 10MP Telephoto",
       "display": "6.7-inch OLED, 120Hz",
       "chipset": "Snapdragon 8 Gen 2",
-      "network_technology": "5G/4G/3G/Edge/GSM"
+      "network_technology": "5G/4G/3G/Edge/GSM",
+      "image": "/images/Galaxy_S24.avif"
     },
     {
       "phone_name": "Ultra Speedy Z",
@@ -16,7 +17,8 @@
       "camera": "50MP Dual Camera",
       "display": "6.5-inch LCD, 90Hz",
       "chipset": "MediaTek Dimensity 920",
-      "network_technology": "5G/4G/3G/GSM"
+      "network_technology": "5G/4G/3G/GSM",
+      "image": "/images/Galaxy_S24.avif"
     },
     {
       "phone_name": "Galaxy Vision 23",
@@ -25,7 +27,8 @@
       "camera": "64MP + 16MP Ultra-wide",
       "display": "6.8-inch AMOLED, 120Hz",
       "chipset": "Exynos 2200",
-      "network_technology": "5G/4G/3G/Edge/GSM"
+      "network_technology": "5G/4G/3G/Edge/GSM",
+      "image": "/images/Galaxy_S24.avif"
     },
     {
       "phone_name": "PixelCraft 7a",
@@ -34,7 +37,8 @@
       "camera": "48MP + 8MP Ultra-wide",
       "display": "6.4-inch OLED, 90Hz",
       "chipset": "Google Tensor G3",
-      "network_technology": "5G/4G/3G"
+      "network_technology": "5G/4G/3G",
+      "image": "/images/Galaxy_S24.avif"
     },
     {
       "phone_name": "Aspire Neo 5",
@@ -43,7 +47,8 @@
       "camera": "64MP + 12MP Macro",
       "display": "6.3-inch AMOLED, 120Hz",
       "chipset": "Snapdragon 778G",
-      "network_technology": "5G/4G/3G"
+      "network_technology": "5G/4G/3G",
+      "image": "/images/Galaxy_S24.avif"
     },
     {
       "phone_name": "Innovate Prime",
@@ -52,7 +57,8 @@
       "camera": "32MP Selfie + 12MP Wide",
       "display": "6.55-inch AMOLED, 120Hz",
       "chipset": "Dimensity 8100",
-      "network_technology": "5G/4G/3G/Edge/GSM"
+      "network_technology": "5G/4G/3G/Edge/GSM",
+      "image": "/images/Galaxy_S24.avif"
     },
     {
       "phone_name": "LiteEdge Alpha",
@@ -61,7 +67,8 @@
       "camera": "20MP + 2MP Macro",
       "display": "5.5-inch LCD, 60Hz",
       "chipset": "Snapdragon 695",
-      "network_technology": "4G/3G/GSM"
+      "network_technology": "4G/3G/GSM",
+      "image": "/images/Galaxy_S24.avif"
     },
     {
       "phone_name": "MegaPhone Z Ultra",
@@ -70,7 +77,8 @@
       "camera": "200MP + 50MP Ultra-wide",
       "display": "7.0-inch AMOLED, 144Hz",
       "chipset": "Snapdragon 8 Gen 3",
-      "network_technology": "5G/4G"
+      "network_technology": "5G/4G",
+      "image": "/images/Galaxy_S24.avif"
     },
     {
       "phone_name": "Quantum Elite",
@@ -79,7 +87,8 @@
       "camera": "120MP + 32MP Ultra-wide",
       "display": "6.75-inch QHD+, 120Hz",
       "chipset": "Kirin 990",
-      "network_technology": "5G/4G/3G/GSM"
+      "network_technology": "5G/4G/3G/GSM",
+      "image": "/images/Galaxy_S24.avif"
     },
     {
       "phone_name": "ThunderBlast 500",
@@ -88,7 +97,8 @@
       "camera": "50MP Quad Camera",
       "display": "6.9-inch AMOLED, 120Hz",
       "chipset": "Exynos 2100",
-      "network_technology": "5G/4G/3G/Edge/GSM"
+      "network_technology": "5G/4G/3G/Edge/GSM",
+      "image": "/images/Galaxy_S24.avif"
     }
   ]
 }
